Added lose condition and lose screen.
Created a script and a cube mesh trigger. When the player hits the mesh their screen will start turning to black (over 5 seconds). When they exit the mesh they go back to normal (2 seconds). If they stay for the full 5 seconds they lose. I've also made sure that the cursor stays unlocked when on the lose or win screen.